chore(tests): separate double commands
